// src/app/interfaces/actualizarHospitalResponse.ts
// Generated by https://quicktype.io

export interface ActualizarHospitalResponse {
  ok:                  boolean;
  message:             string;
  hospitalActualizado: HospitalActualizado;
}

export interface HospitalActualizado {
  _id:     string;
  nombre:  string;
  usuario: string;
  img:     string;
}