// src/app/interfaces/getHospitalesResponse.ts
// Generated by https://quicktype.io

import { HospitalUsuario } from "../models/hospital.model";

// Generated by https://quicktype.io

export interface GetHospitalesResponse {
  ok:             boolean;
  hospitales:     Hospital[];
  totalRegistros: number;
}

export interface Hospital {
  _id:      string;
  nombre:   string;
  usuario?: HospitalUsuario;
  img:      string;
  email?:   string;
}